% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objects.R
\name{get_objects}
\alias{get_objects}
\title{Get objects from the RATO ArcGIS Enterprise environment via the API}
\usage{
get_objects(object_ids, token = get_token(), batch_size = 50)
}
\arguments{
\item{object_ids}{Vector of object_ids, will be converted to character.}

\item{token}{Access token, by default generated by \code{get_token()}}

\item{batch_size}{Number of objects to request per API call, default is 50.
Setting this to a lower number will result in more API calls, but will also
reduce the risk of timeouts or errors when requesting large datasets. When
set to a higher number, the API may timeout or return an error if the
dataset is too large.}
}
\value{
tibble of requested objects.
}
\description{
This function retrieves objects from the RATO ArcGIS environment based on
the object_id. It is capable of retrieving a large amount of objects in a
single function call. As to minimize errors it will  fetch \verb{batch size}
number of objects per request, and will operate in parallel: for a maximum of
5000 requests per 60 seconds, having at most 10 requests open at a time.
}
